refactor: extract guard SQL, remove dead code, fix notifications (Batch I)
SQL extraction:
- Move GuardBlockSummary + query from notifications to analytics.DB method
- No raw SQL referencing analytics tables remains in notifications

Dead code:
- Delete WriteFeedCacheJSON/WriteFeedCacheJSONTo (zero production callers)
- Delete 3 associated tests

CodeRabbit fixes:
- Narrow ALTER TABLE error handling to only ignore duplicate column errors
- Mark expired notifications as surfaced to prevent infinite rescan
